feat: Add refactor
