+++ App.tsx
@@ -1,28 +1,12 @@
 import { StyleSheet, SafeAreaView, Platform } from 'react-native'
 import { StatusBar } from 'expo-status-bar'
-import { NavigationContainer } from '@react-navigation/native'
-import { createNativeStackNavigator } from '@react-navigation/native-stack'
-import { RootStackRiackAndMortyParams } from './src/shared'
-import { RickAndMortyApp } from './src/screens/RickAndMortyApp'
-import { CharacterDetails } from './src/screens/CharacterDetails'
-
-const Stack = createNativeStackNavigator<RootStackRiackAndMortyParams>()
-
-const routeScreenDefaultOptions = {
-  headerStyle: { backgroundColor: '#01aeca' },
-  headerTitleStyle: { color: '#fff' }
-}
+import { RickAndMortyRouter } from './src/router/RiackAndMortyRouter'
 
 const App: React.FC = () => {
 
   return (
     <SafeAreaView style={styles.container}>
-      <NavigationContainer>
-        <Stack.Navigator initialRouteName='Home'>
-          <Stack.Screen name='Home' component={RickAndMortyApp} />
-          <Stack.Screen name='Details' options={{ ...routeScreenDefaultOptions, title: 'Rick and Morty' }} component={CharacterDetails} />
-        </Stack.Navigator>
-      </NavigationContainer>
+      <RickAndMortyRouter />
       <StatusBar style='auto' />
     </SafeAreaView>
   )
